refactor: useGetData 경로 수정
